% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cper.r
\name{cmat}
\alias{cmat}
\title{data frame wrapper for caret::confusionMatrix}
\usage{
cmat(df, posclass = NULL)
}
\arguments{
\item{df}{A data frame with first column the data, second column the reference}

\item{posclass}{(optional) output name of positive class}
}
\description{
data frame wrapper for caret::confusionMatrix
}
\examples{
df \%>\% select(prediction, ref) \%>\% cmat()
}
