% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.r
\name{cv2logsd}
\alias{cv2logsd}
\title{Convert CV on original scale to sd on log scale}
\usage{
cv2logsd(cv, base = 2)
}
\arguments{
\item{cv}{cv on original scale}

\item{logbase}{base of the log scale (default = 2)}
}
\description{
Convert CV on original scale to sd on log scale
}
