% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.r
\name{logsd2cv}
\alias{logsd2cv}
\title{Convert sd on log scale to CV on original scale}
\usage{
logsd2cv(s, base = 2)
}
\arguments{
\item{s}{standard deviation on log scale}

\item{logbase}{base of the log scale (default = 2)}
}
\description{
Convert sd on log scale to CV on original scale
}
