% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cper.r
\name{perfci}
\alias{perfci}
\title{performance metrics with CI}
\usage{
perfci(cmat)
}
\arguments{
\item{cmat}{confusionMatrix object as returned by caret::confusionMatrix}
}
\description{
performance metrics with CI
}
