% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.r
\name{read.crf.sheet}
\alias{read.crf.sheet}
\title{read sheetv from excel file of Avania CRF}
\usage{
read.crf.sheet(dbFile, sheet)
}
\arguments{
\item{dbFile}{file name of the export}

\item{sheet}{str identifying the table to import}
}
\description{
read sheetv from excel file of Avania CRF
}
