% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.r
\name{read.crf.table}
\alias{read.crf.table}
\title{read tab file of Avania CRF}
\usage{
read.crf.table(dbExport, tabstr, ext = ".txt", csep = "\\t")
}
\arguments{
\item{tabstr}{str identifying the table to import}

\item{ext}{extension}

\item{csep}{separator}

\item{dbexport}{base name of the export}
}
\description{
read tab file of Avania CRF
}
