% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.r
\name{write.delim}
\alias{write.delim}
\title{write table to tab delimited file}
\usage{
write.delim(df, filename)
}
\arguments{
\item{df}{A data frame}

\item{filename}{output file name}
}
\description{
write table to tab delimited file
}
\examples{
write.delim(iris, filename = "iris.txt")
}
